% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EphData.R
\name{EphData}
\alias{EphData}
\title{Eph Data}
\usage{
EphData(dataset)
}
\arguments{
\item{dataset}{The dataset that the function will read and parse. The options are
"names20xx" where xx can be any number from 04 to 16.}
}
\value{
An 8 column dataframe containing information about Williams College graduates. In order
by column, the dataframe contains year, what year the student graduted, latin honors, which latin
honors, if any, did the student earn, first name, first name length, last name, last name length,
subject honors, highest for highest honors, with for honors, and blank for no honors, and subect
those honors were earned, blank if no honors were earned.
}
\description{
Reads information about Williams College Graduating classes between 2004 and 2016
and constructs a dataset containing the information the college provides
}

