% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameData.R
\docType{data}
\name{nameData}
\alias{nameData}
\title{Information about Williams College graduates 2004 to 2016}
\format{A data frame with 6875 rows and 9 variables:
\describe{
  \item{year}{year}
  \item{Latin Honors}{Latin Honors earned by students, summa, magna, cum, or blank for none}
  \item{first name}{Student first name}
  \item{length}{Number of characters in first name}
  \item{last name}{Student last name}
  \item{length}{Number of characters in last name}
  \item{subject honors}{Subject honors earned, highest for highest honors, with for honors, and blank for no honors}
  \item{subject}{Subject in which the study earned honors, black if no honors}
  \item{Asian}{Whether or not a student is estiamted to be Asian, 1 for Asian, 0 for not Asian}
  ...
}}
\source{
\url{http://web.williams.edu/admin/registrar/catalog/archive.html}
}
\usage{
nameData
}
\description{
An 9 column dataframe containing information about Williams College graduates.
}
\keyword{datasets}

