% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSummary.R
\name{nameSummary}
\alias{nameSummary}
\title{Name Summary}
\usage{
nameSummary(type)
}
\arguments{
\item{type}{The type of summary statistic or graphic the function will produce. The options
are honors, length, and popular.}
}
\value{
If the user codes honors, the function will return a bar graph showing the proportion of
Asian and non-Asian students that earn Latin and Subject honors. If the user codes length, the
function will return a bar graph comparing the lengths of the first and last names of Asian
and non-Asian students. If the user codes popular, the function will return a bar graph showing
the correlation betweeen the lengths of first and last names, and the proportion of the single most
popular name between 2004 and 2016.
}
\description{
Produces summary statistics and graphics for Williams College graduates between
2004 and 2016. The analysis is focused on the differenced between students that are guessed
to be Asian versus students who are guessed not to be Asian. Surname analysis is employed to make
this decision. The surnames are compared with 378 common Asian surnames. The positive predictive
rate is estimated to be 21%.
}

