% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot Region}
\usage{
plot_region(region)
}
\arguments{
\item{region}{The region that will be plotted. The possible options are West, Mountain West, South, Midwest, Northeast,
 Mid Atlantic, and All.
They should be coded as strings without spaces and with capital letter where normal. For example, /code{"MidAtlantic"}}
}
\value{
A barplot for the specified region. If the user codes /code{"All"}, the function will return all regions.
}
\description{
Uses the geographic distribution data from the read function to generate graphics for user coded
regions
}

