% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot State}
\usage{
plot_state(state)
}
\arguments{
\item{region}{The region that will be plotted. The possible options are West, Mountain West, South, Midwest, Northeast, Mid Atlantic, and All.
They should be coded as strings without spaces and with capital letter where normal. For example, /code{"NorthCarolina"}}
}
\value{
A barplot for the specified region. If the user codes /code{"Notable"}, the function will return notable states.
}
\description{
Uses the geographic distribution data from the readEphData function to generate graphics for user coded states.
}

