% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNameData.R
\name{readNameData}
\alias{readNameData}
\title{Read Name Data}
\usage{
readNameData(dataset)
}
\arguments{
\item{dataset}{The dataset that the function will read and parse. The options are "names20xx" where xx can be any number from 04 to 16.}
}
\value{
A dataframe containing the first names and length of first names for all Williams College Graduates of a certain year.
}
\description{
Reads the first names of Williams College graduating classes from 2004 to 2016
}

