% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{summary}
\alias{summary}
\title{Summary}
\usage{
summary(type)
}
\arguments{
\item{type}{The type of statistics to be computed. The options are "popular", "most popular", "name length",
"name length plot", and "least popular."}
}
\value{
If the user codes popular, the function returns a dataframe with the six most popular names
over the past 13 years. If the user codes most popular, the function returns a dataframe reporting all
names that have been in the top six over the past 13 years, and how many times each name has been in the top six.
If the user codes name length, the function returns a dataframe reporting the average first name length of the
graduating class for each year. If the user codes name length plot, the function reports name length data in
the form of a scatterplot. If the user codes least popular, the function returns a random sample of first names that
have only occurred at Williams one time since 2004.
}
\description{
Uses the data from the readNameData() function to compute relavent summary statistics
}

