% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totaldistribution.R
\name{total_dist}
\alias{total_dist}
\title{Total Dist}
\usage{
total_dist(type)
}
\arguments{
\item{type}{The type of output generated. The possible options are Summary, US, and SD. Summary generates
a dataframe containing the geographic distribution for the years 2001 to 2016. US generates a barplot displaying
how total United States enrollment has changed at Williams from 2001 to 2016. SD generates a data frame containing
the standard deviation in total United States enrollment for the year blocks 2001-2010 and 2011-2016.
For example, /code{"Summary"}}
}
\value{
Dataframe or barplot depending on type.
}
\description{
Uses the geographic distribution data from the readEphData function to generate graphics and summary statistic for user coded options.
}

