% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state.R
\name{plot_state}
\alias{plot_state}
\title{Plot State}
\usage{
plot_state(state)
}
\arguments{
\item{state}{The state that will be plotted. The possible options are all 50 states, Guam, Puerto Rico,
District of Columbia, and Notable. They should be coded as strings without spaces and with capital letters
where normal. For example, /code{"NorthCarolina"}}
}
\value{
A barplot for the specified state. If the user codes /code{"Notable"}, the function will return notable states.
}
\description{
Uses the geographic distribution data from the readEphData function to generate graphics for user coded states.
}

