% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEphData.R
\name{readEphData}
\alias{readEphData}
\title{readEphData}
\usage{
readEphData()
}
\value{
Data frame containing the number of students from each state or province from 2001-2016.
/code{} automatically returns data from all states and provinces and from all years between 2001 and 2016.
The user can easily subset if a smaller dataset is desired.
}
\description{
Scan data on the United States geographic distribution of Williams Students from 2001 to 2016. It includes
all 50 states, District of Columbia, Guam, and Puerto Rico. It excludes students from the Armed Forces Europe or
Pacific and the Virgin Islands.
}

